Refactor: PostDetailComponent 즐겨찾기 아이콘 컴포넌트로 분리
PostDetailBookmarkIcon 제작하여 분리 후  PostDetailComponent에 적용
